move override code into own file
